pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-poc"
version = "0.1.0"
description = "Model Context Protocol Proof of Concept"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "llama-index>=0.10.0",
    "llama-index-llms-openai>=0.1.0",
    "llama-index-tools-mcp>=0.1.0",
    "llama-index-core>=0.10.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
] 